requirements.txt
Django==5.0.1
django-filter==23.5
djangorestframework==3.14.0

